Fix dynamic loading functions of libnetsys_client.z.so by saving them globally
Before this fix, every call to libnetsys_client.z.so's functions involves
loading the lib, searching symbols and unloading the lib. Assuming debugging a
process with lldb, it will hurt the debugging performance because lldb stops at
dlopen and dlcose (lldb inserts internal breakpoints to know which shared
library is loaded or unloaded). It is fixed by storing the handle to the lib and
function pointers globally.

Issue: I6XDBW
Test: passed most test cases of libc-test with 4 anticipated cases failed
Signed-off-by: hongbinj <jinhongbin2@huawei.com>