Add SetThreadInfoCallback to libc.map.txt
SetThreadInfoCallback is introduced by dfx_signal_handler which is static linked to libc.
If you want to record some crash-related thread running information when the crash signal arrives.
You can use this mechanism to register callbacks. This callback is called in the signal handler,
so you must handle it with care. You can store additional information in the passed buffer,
which will be displayed in the FatalMessage field of the crash log.

Signed-off-by: MapleStory <zengzhi5@huawei.com>